Add support for VS2019.
