+++ omaha/base/string.h
@@ -18,6 +18,7 @@
 
 #include <windows.h>
 #include <stdint.h>
+
 #include <limits>
 #include <string>
 #include <vector>
