+++ omaha/omaha_version_utils.py
@@ -95,9 +95,9 @@ VC90  = 1500  # VC2008/VC90 (not supported by the current build).
 VC100 = 1600  # VC2010/VC10
 VC110 = 1700  # VC2012/VC11 (not supported by the current build).
 VC120 = 1800  # VC2013/VC12
-VC140 = 1900  # VC2015/VC14.0
+VC140 = 1900  # VC2015/VC14
 VC150 = 1910  # VC2017 version 15.0-15.9 / VC14.1-14.16
-VC160 = 1920  # VS2019 version 16.0 / VC14.2
+VC160 = 1920  # VC2019 version 16.0 / VC14.2
 
 def _IsSupportedOmaha2Version(omaha_version):
   """Returns true if omaha_version is an Omaha 2 version and is supported."""
