+++ omaha/site_scons/site_tools/omaha_builders.py
@@ -246,14 +246,14 @@ def ConfigureEnvFor64Bit(env):
                                    '$ATLMFC_VC14_0_DIR/lib/amd64',
                                    platform_sdk_lib_dir + '/um/x64',
                                    platform_sdk_lib_dir + '/ucrt/x64',],
-      omaha_version_utils.VC150: [ '$VC15_0_DIR' + '/lib/x64',
+      omaha_version_utils.VC150: [ '$VC15_0_DIR/lib/x64',
                                    '$ATLMFC_VC15_0_DIR/lib/x64',
-                                   '$WINDOWS_SDK_10_0_LIB_DIR' + '/um/x64',
-                                   '$WINDOWS_SDK_10_0_LIB_DIR' + '/ucrt/x64',],
+                                   '$WINDOWS_SDK_10_0_LIB_DIR/um/x64',
+                                   '$WINDOWS_SDK_10_0_LIB_DIR/ucrt/x64',],
       omaha_version_utils.VC160: [ '$VC16_0_DIR/lib/x64',
                                    '$ATLMFC_VC16_0_DIR/lib/x64',
-                                   platform_sdk_lib_dir + '/um/x64',
-                                   platform_sdk_lib_dir + '/ucrt/x64',],
+                                   '$WINDOWS_SDK_10_0_LIB_DIR/um/x64',
+                                   '$WINDOWS_SDK_10_0_LIB_DIR/ucrt/x64',],
       }[env['msc_ver']]
 
   env.Prepend(LIBPATH=lib_paths)
@@ -264,7 +264,7 @@ def ConfigureEnvFor64Bit(env):
        omaha_version_utils.VC100 : '$VC10_0_DIR/vc/bin/x86_amd64',
        omaha_version_utils.VC120 : '$VC12_0_DIR/vc/bin/x86_amd64',
        omaha_version_utils.VC140 : '$VC14_0_DIR/vc/bin/x86_amd64',
-       omaha_version_utils.VC150 : '$VC15_0_DIR' + '/bin/HostX64/x64',
+       omaha_version_utils.VC150 : '$VC15_0_DIR/bin/HostX64/x64',
        omaha_version_utils.VC160 : '$VC16_0_DIR/bin/HostX64/x64'}
       [env['msc_ver']]))
